feat(tmux): show hostname in status if present
just so I have some visual feedback that I'm on a server.
Also testing how having the session_path in place of the session name
feels like - in exchange for the clock (which can be invoked via c-a t)
and the current pane indicator, which I never really needed. To keep it
simple.